% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{billingAccounts.sinks.get}
\alias{billingAccounts.sinks.get}
\title{Gets a sink.}
\usage{
billingAccounts.sinks.get(sinkName)
}
\arguments{
\item{sinkName}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/logging.admin
\item https://www.googleapis.com/auth/logging.read
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/logging.admin, https://www.googleapis.com/auth/logging.read)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/logging/docs/}{Google Documentation}
}
