% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{billingAccounts.sinks.update}
\alias{billingAccounts.sinks.update}
\title{Updates a sink. If the named sink doesn't exist, then this method is identical to sinks.create. If the named sink does exist, then this method replaces the following fields in the existing sink with values from the new sink: destination, filter, output_version_format, start_time, and end_time. The updated filter might also have a new writer_identity; see the unique_writer_identity field.}
\usage{
billingAccounts.sinks.update(LogSink, sinkName, uniqueWriterIdentity = NULL)
}
\arguments{
\item{LogSink}{The \link{LogSink} object to pass to this method}

\item{sinkName}{Required}

\item{uniqueWriterIdentity}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/logging.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/logging.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/logging/docs/}{Google Documentation}

Other LogSink functions: \code{\link{LogSink}},
  \code{\link{billingAccounts.sinks.create}},
  \code{\link{folders.sinks.create}},
  \code{\link{folders.sinks.update}},
  \code{\link{organizations.sinks.create}},
  \code{\link{organizations.sinks.update}},
  \code{\link{projects.sinks.create}},
  \code{\link{projects.sinks.update}}
}
