% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{folders.sinks.create}
\alias{folders.sinks.create}
\title{Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the current time is outside the sink's start and end times or the sink's writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.}
\usage{
folders.sinks.create(LogSink, parent, uniqueWriterIdentity = NULL)
}
\arguments{
\item{LogSink}{The \link{LogSink} object to pass to this method}

\item{parent}{Required}

\item{uniqueWriterIdentity}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/logging.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/logging.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/logging/docs/}{Google Documentation}

Other LogSink functions: \code{\link{LogSink}},
  \code{\link{billingAccounts.sinks.create}},
  \code{\link{billingAccounts.sinks.update}},
  \code{\link{folders.sinks.update}},
  \code{\link{organizations.sinks.create}},
  \code{\link{organizations.sinks.update}},
  \code{\link{projects.sinks.create}},
  \code{\link{projects.sinks.update}}
}
