% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{HttpRequest}
\alias{HttpRequest}
\title{HttpRequest Object}
\usage{
HttpRequest(latency = NULL, userAgent = NULL, cacheFillBytes = NULL,
  requestMethod = NULL, requestSize = NULL, responseSize = NULL,
  requestUrl = NULL, remoteIp = NULL, serverIp = NULL,
  cacheLookup = NULL, cacheHit = NULL,
  cacheValidatedWithOriginServer = NULL, status = NULL, referer = NULL)
}
\arguments{
\item{latency}{The request processing latency on the server, from the time the request was received until the response was sent}

\item{userAgent}{The user agent sent by the client}

\item{cacheFillBytes}{The number of HTTP response bytes inserted into cache}

\item{requestMethod}{The request method}

\item{requestSize}{The size of the HTTP request message in bytes, including the request headers and the request body}

\item{responseSize}{The size of the HTTP response message sent back to the client, in bytes, including the response headers and the response body}

\item{requestUrl}{The scheme (http, https), the host name, the path and the query portion of the URL that was requested}

\item{remoteIp}{The IP address (IPv4 or IPv6) of the client that issued the HTTP request}

\item{serverIp}{The IP address (IPv4 or IPv6) of the origin server that the request was sent to}

\item{cacheLookup}{Whether or not a cache lookup was attempted}

\item{cacheHit}{Whether or not an entity was served from cache (with or without validation)}

\item{cacheValidatedWithOriginServer}{Whether or not the response was validated with the origin server before being served from cache}

\item{status}{The response code indicating the status of response}

\item{referer}{The referer URL of the request, as defined in HTTP/1}
}
\value{
HttpRequest object
}
\description{
HttpRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A common proto for logging HTTP requests. Only contains semantics defined by the HTTP specification. Product-specific logging information MUST be defined in a separate message.
}
