% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListLogEntriesRequest}
\alias{ListLogEntriesRequest}
\title{Stackdriver Logging API Objects 
Writes log entries and manages your Stackdriver Logging configuration.}
\usage{
ListLogEntriesRequest(projectIds = NULL, filter = NULL, pageToken = NULL,
  pageSize = NULL, orderBy = NULL, resourceNames = NULL)
}
\arguments{
\item{projectIds}{Deprecated}

\item{filter}{Optional}

\item{pageToken}{Optional}

\item{pageSize}{Optional}

\item{orderBy}{Optional}

\item{resourceNames}{Required}
}
\value{
ListLogEntriesRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:01:01
filename: /Users/mark/dev/R/autoGoogleAPI/googleloggingv2.auto/R/logging_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListLogEntriesRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The parameters to ListLogEntries.
}
\seealso{
Other ListLogEntriesRequest functions: \code{\link{entries.list}}
}
