% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListLogEntriesResponse}
\alias{ListLogEntriesResponse}
\title{ListLogEntriesResponse Object}
\usage{
ListLogEntriesResponse(entries = NULL, nextPageToken = NULL)
}
\arguments{
\item{entries}{A list of log entries}

\item{nextPageToken}{If there might be more results than those appearing in this response, then nextPageToken is included}
}
\value{
ListLogEntriesResponse object
}
\description{
ListLogEntriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result returned from ListLogEntries.
}
