% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListLogsResponse}
\alias{ListLogsResponse}
\title{ListLogsResponse Object}
\usage{
ListLogsResponse(nextPageToken = NULL, logNames = NULL)
}
\arguments{
\item{nextPageToken}{If there might be more results than those appearing in this response, then nextPageToken is included}

\item{logNames}{A list of log names}
}
\value{
ListLogsResponse object
}
\description{
ListLogsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result returned from ListLogs.
}
