% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListSinksResponse}
\alias{ListSinksResponse}
\title{ListSinksResponse Object}
\usage{
ListSinksResponse(nextPageToken = NULL, sinks = NULL)
}
\arguments{
\item{nextPageToken}{If there might be more results than appear in this response, then nextPageToken is included}

\item{sinks}{A list of sinks}
}
\value{
ListSinksResponse object
}
\description{
ListSinksResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result returned from ListSinks.
}
