% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogEntry.jsonPayload}
\alias{LogEntry.jsonPayload}
\title{LogEntry.jsonPayload Object}
\usage{
LogEntry.jsonPayload()
}
\value{
LogEntry.jsonPayload object
}
\description{
LogEntry.jsonPayload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The log entry payload, represented as a structure that is expressed as a JSON object.
}
\seealso{
Other LogEntry functions: \code{\link{LogEntry.labels}},
  \code{\link{LogEntry.protoPayload}},
  \code{\link{LogEntry}}
}
