% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogEntry.protoPayload}
\alias{LogEntry.protoPayload}
\title{LogEntry.protoPayload Object}
\usage{
LogEntry.protoPayload()
}
\value{
LogEntry.protoPayload object
}
\description{
LogEntry.protoPayload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The log entry payload, represented as a protocol buffer. Some Google Cloud Platform services use this field for their log entry payloads.
}
\seealso{
Other LogEntry functions: \code{\link{LogEntry.jsonPayload}},
  \code{\link{LogEntry.labels}}, \code{\link{LogEntry}}
}
