% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogEntrySourceLocation}
\alias{LogEntrySourceLocation}
\title{LogEntrySourceLocation Object}
\usage{
LogEntrySourceLocation(file = NULL, function. = NULL, line = NULL)
}
\arguments{
\item{file}{Optional}

\item{function.}{Optional}

\item{line}{Optional}
}
\value{
LogEntrySourceLocation object
}
\description{
LogEntrySourceLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional information about the source code location that produced the log entry.
}
