% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogMetric}
\alias{LogMetric}
\title{LogMetric Object}
\usage{
LogMetric(description = NULL, version = NULL, filter = NULL,
  name = NULL)
}
\arguments{
\item{description}{Optional}

\item{version}{Output only}

\item{filter}{Required}

\item{name}{Required}
}
\value{
LogMetric object
}
\description{
LogMetric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a logs-based metric. The value of the metric is the number of log entries that match a logs filter in a given time interval.
}
\seealso{
Other LogMetric functions: \code{\link{projects.metrics.create}},
  \code{\link{projects.metrics.update}}
}
