% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{MonitoredResource.labels}
\alias{MonitoredResource.labels}
\title{MonitoredResource.labels Object}
\usage{
MonitoredResource.labels()
}
\value{
MonitoredResource.labels object
}
\description{
MonitoredResource.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Required. Values for all of the labels listed in the associated monitored resource descriptor. For example, Cloud SQL databases use the labels 'database_id' and 'zone'.
}
\seealso{
Other MonitoredResource functions: \code{\link{MonitoredResource}}
}
