% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{projects.metrics.create}
\alias{projects.metrics.create}
\title{Creates a logs-based metric.}
\usage{
projects.metrics.create(LogMetric, parent)
}
\arguments{
\item{LogMetric}{The \link{LogMetric} object to pass to this method}

\item{parent}{The resource name of the project in which to create the metric:}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/logging.admin
\item https://www.googleapis.com/auth/logging.write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/logging.admin, https://www.googleapis.com/auth/logging.write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/logging/docs/}{Google Documentation}

Other LogMetric functions: \code{\link{LogMetric}},
  \code{\link{projects.metrics.update}}
}
