% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{projects.metrics.delete}
\alias{projects.metrics.delete}
\title{Deletes a logs-based metric.}
\usage{
projects.metrics.delete(metricName)
}
\arguments{
\item{metricName}{The resource name of the metric to delete:}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/logging.admin
\item https://www.googleapis.com/auth/logging.write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/logging.admin, https://www.googleapis.com/auth/logging.write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/logging/docs/}{Google Documentation}
}
