% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{projects.sinks.delete}
\alias{projects.sinks.delete}
\title{Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.}
\usage{
projects.sinks.delete(sinkName)
}
\arguments{
\item{sinkName}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/logging.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/logging.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/logging/docs/}{Google Documentation}
}
