% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{WriteLogEntriesRequest.labels}
\alias{WriteLogEntriesRequest.labels}
\title{WriteLogEntriesRequest.labels Object}
\usage{
WriteLogEntriesRequest.labels()
}
\value{
WriteLogEntriesRequest.labels object
}
\description{
WriteLogEntriesRequest.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional. Default labels that are added to the labels field of all log entries in entries. If a log entry already has a label with the same key as a label in this parameter, then the log entry's label is not changed. See LogEntry.
}
\seealso{
Other WriteLogEntriesRequest functions: \code{\link{WriteLogEntriesRequest}},
  \code{\link{entries.write}}
}
